% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAPIKey.R
\name{addAPIKey}
\alias{addAPIKey}
\title{addAPIKey}
\usage{
addAPIKey(conn, user_id, user_email)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{user_id}{A list of user_ids in the database}
}
\description{
Add an API Key for one or more users in the database
}
