% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAssayPlatform.R
\name{addAssayPlatform}
\alias{addAssayPlatform}
\title{addAssayPlatform}
\usage{
addAssayPlatform(accession, name, species, handle)
}
\arguments{
\item{accession}{string: accession ID of the platform}

\item{name}{string: name of the platform}

\item{species}{string: species that the platform is used on}

\item{handle}{connection handle object. ensure the user has insert
privilages. see ?newConnHandle for more information}
}
\description{
inserts an assay platform into the assay_platforms table
in your SigRepo database.
}
