% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCollectionEntity.R
\name{addCollectionEntity}
\alias{addCollectionEntity}
\title{addCollectionEntity}
\usage{
addCollectionEntity(
  connHandle,
  collectionName = NULL,
  collectionDescription = NULL
)
}
\arguments{
\item{connHandle}{connection handle to use for query}

\item{collectionName}{name of the collection you're adding}

\item{collectionDescription}{description of the collection you're adding}
}
\description{
adds signature Collection entities to signature_collections
table in the Database.
}
