% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCollectionSignatures.R
\name{addCollectionSignatures}
\alias{addCollectionSignatures}
\title{addCollectionSignatures}
\usage{
addCollectionSignatures(thisHandle, collectionName = NULL, signatures = NULL)
}
\arguments{
\item{thisHandle}{the connection handle used for executing the insert/s}

\item{collectionName}{character vector denoting the collection}

\item{signatures}{character vector of signature names to associate with
collectionName parameter}
}
\description{
adds signature-collection pairs to signatures_to_collections
table in the database.
}
