% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{addFeatures}
\alias{addFeatures}
\title{addFeatures}
\usage{
addFeatures(featureName, referenceGenome, featureType, handle)
}
\arguments{
\item{featureName}{string: name of the feature}

\item{referenceGenome}{string: name of the reference genome}

\item{featureType}{string: name of the feature type e.g. "gene"}

\item{handle}{your connection handle. ensure you have insert privilages on
your system}
}
\description{
inserts features into a database. Ensure that the user doing so
has insert privilages.
}
