% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKeywordSignatures.R
\name{addKeywordSignatures}
\alias{addKeywordSignatures}
\title{addKeywordSignatures}
\usage{
addKeywordSignatures(thisHandle, keywords = NULL, signature = NULL)
}
\arguments{
\item{thisHandle}{the connection handle used for executing the insert/s}

\item{keywords}{character vector denoting the keywords with which to
associate a signature}

\item{signature}{character vector of signature name to associate with
keywords parameter}
}
\description{
adds signature-keyword pairs to signatures_to_collections
table in the database.
}
