% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLevel2.R
\name{addLevel2}
\alias{addLevel2}
\title{addLevel2 data}
\usage{
addLevel2(lv2_file, sid, sig_name, thisHandle, verbose = F)
}
\arguments{
\item{lv2_file}{file path to lvl2/3 data file,}

\item{sid}{signature id,}

\item{sig_name}{signature_name}

\item{thisHandle}{connection handle to use}

\item{verbose}{print out messages as the function executes. False by default}
}
\description{
function will insert lvl2/3 data into DB
}
