% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOrganism.R
\name{addOrganism}
\alias{addOrganism}
\title{addOrganism}
\usage{
addOrganism(conn_handler, organism_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{organism_tbl}{A data frame object contains appropriate column names: 
organism (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add organisms to database
}
