% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhenotype.R
\name{addPhenotype}
\alias{addPhenotype}
\title{addPhenotype}
\usage{
addPhenotype(conn, phenotype_tbl)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{phenotype_tbl}{An data frame containing appropriate column names: phenotype}
}
\description{
Add phenotype to database
}
