% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlatform.R
\name{addPlatform}
\alias{addPlatform}
\title{addPlatform}
\usage{
addPlatform(conn_handler, platform_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{platform_tbl}{An data frame containing appropriate column names: 
platform_name (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add platforms to database
}
