% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProteomicsFeatureSet.R
\name{addProteomicsFeatureSet}
\alias{addProteomicsFeatureSet}
\title{addProteomicsFeatureSet}
\usage{
addProteomicsFeatureSet(conn_handler, feature_set, verbose = TRUE)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{feature_set}{A data frame containing appropriate column names: 
feature_name, organism, description, synonyms, n_synonyms, ensemble_ids, 
n_ensemble_ids, transcript_biotypes, chromosome_name, start_position, 
end_position}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add Proteomics Feature Set into database
}
\keyword{internal}
