% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRefFeatureSet.R
\name{addRefFeatureSet}
\alias{addRefFeatureSet}
\title{addRefFeatureSet}
\usage{
addRefFeatureSet(
  conn_handler,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  feature_set,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{assay_type}{Type of assays: transcriptomics, proteomics, metabolomics, 
methylomics, genetic_variations, dna_binding_sites}

\item{feature_set}{A data frame containing appropriate column names}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add reference feature set into database
}
