% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleType.R
\name{addSampleType}
\alias{addSampleType}
\title{addSampleType}
\usage{
addSampleType(conn_handler, sample_type_tbl)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{sample_type_tbl}{An data frame containing appropriate column names:
sample_type, brenda_accession}
}
\description{
Add sample_type to database
}
