% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureAccess.R
\name{addSignatureAccess}
\alias{addSignatureAccess}
\title{addSignatureAccess}
\usage{
addSignatureAccess(conn, access_tbl)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{access_tbl}{A data frame containing the appropriate column names:
signature_id, user_id, access_type}
}
\description{
Add signature access information to database
}
