% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCollection.R, R/addSignatureCollection.R
\name{addSignatureCollection}
\alias{addSignatureCollection}
\title{addSignatureCollection}
\usage{
addSignatureCollection(conn_handler, omic_collection)

addSignatureCollection(conn_handler, omic_collection)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{omic_collection}{A collection of OmicSignature objects from OmicSignature package}
}
\description{
Add signature collection to database

Add signature collection to database
}
