% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureHandler.R
\name{addSignatureHandler}
\alias{addSignatureHandler}
\title{addSignatureHandler}
\usage{
addSignatureHandler(conn, omic_signature)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{omic_signature}{An R6 class object from OmicSignature package}
}
\description{
Add signature to database
}
