% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureKeywords.R
\name{addSignatureKeywords}
\alias{addSignatureKeywords}
\title{addSignatureKeywords}
\usage{
addSignatureKeywords(keyword_v, sid, keywordHandle)
}
\arguments{
\item{keyword_v}{character vector containing keywords}

\item{sid}{signature id associated with keywords}

\item{keywordHandle}{connection handle for keyword-signature insertion to Database}
}
\description{
function for adding keyword-signature pairs
}
