% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureToCollection.R
\name{addSignatureToCollection}
\alias{addSignatureToCollection}
\title{addSignatureToCollection}
\usage{
addSignatureToCollection(conn_handler, collection_id, signature_id)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to
a remote database obtained from SigRepo::newConnhandler()}

\item{collection_id}{ID of the collection in the database}

\item{signature_id}{ID of the signature to be added to the collection}
}
\description{
Add signature to collection table in the database
}
