% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureToQueue.R
\name{addSignatureToQueue}
\alias{addSignatureToQueue}
\title{addSignatureToQueue}
\usage{
addSignatureToQueue(
  signatureName,
  signatureObject,
  verdict = NULL,
  verbose = F
)
}
\arguments{
\item{signatureName}{name of the signature}

\item{signatureObject}{an OmicSignature Object.}

\item{verdict}{whether to add a signature to your system, or to remove it
from your queue}

\item{verbose}{verbosity of the command}
}
\description{
addSignatureToQueue
}
