% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureWithID.R
\name{addSignatureWithID}
\alias{addSignatureWithID}
\title{addSignatureWithID}
\usage{
addSignatureWithID(
  conn_handler,
  omic_signature,
  assign_signature_id,
  assign_user_name,
  visibility = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{omic_signature}{An R6 class object from OmicSignature package}

\item{assign_signature_id}{Assign an unique ID to the uploaded signature.}

\item{assign_user_name}{Assign an unique user name to the uploaded signature.}

\item{visibility}{A logical value indicates whether or not to allow others
to view and access one's uploaded signature. Default is \code{FALSE}.}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{FALSE}.}
}
\description{
Add signature to database with an assigned ID
}
\keyword{internal}
