% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureWrapper.R
\name{addSignatureWrapper}
\alias{addSignatureWrapper}
\title{addSignatureWrapper}
\usage{
addSignatureWrapper(objectFile, thisHandle, uploadPath, thisUser)
}
\arguments{
\item{objectFile}{a json or rds file, formatted in the way of an
OmicSignature Object. It can also be a variable that is the object itself.}

\item{thisHandle}{Database connection handle}

\item{uploadPath}{where to upload the object file}

\item{thisUser}{the username of the submitter}
}
\description{
Function that executes all required inserts and uploads of a signature
}
