% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscriptomicsSignatureSet.R
\name{addTranscriptomicsSignatureSet}
\alias{addTranscriptomicsSignatureSet}
\title{addTranscriptomicsSignatureSet}
\usage{
addTranscriptomicsSignatureSet(
  conn_handler,
  signature_id,
  organism_id,
  signature_set,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{signature_id}{A signature name}

\item{organism_id}{An organism}

\item{signature_set}{A data frame containing the appropriate column names:
feature_name, probe_id, score, direction}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add signature feature set to database
}
\keyword{internal}
