% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUserToCollection.R
\name{addUserToCollection}
\alias{addUserToCollection}
\title{addUserToCollection}
\usage{
addUserToCollection(
  conn_handler,
  collection_id,
  user_name,
  access_type = c("owner", "editor", "viewer")
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to
a remote database obtained from SigRepo::newConnhandler()}

\item{collection_id}{ID of the signature in the database}

\item{user_name}{user name to be added to the signature}

\item{access_type}{give permission to users to access the signature 
in the database as admin/owner/editor/viewer

\code{admin} has read and write access to all signatures
\code{owner} has read and write access to their own uploaded signatures
\code{editor} has read and write access to only the signatures that other 
users were given them access to
\code{viewer} has only read access to the only the signatures that other 
users were given them access to}
}
\description{
Add user to signature access table in the database
}
