% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureAccess.R
\name{addUserToSignature}
\alias{addUserToSignature}
\title{addSignatureAccess}
\usage{
addUserToSignature(
  conn,
  signature_id,
  user_id,
  access_type = c("admin", "owner", "viewer")
)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{signature_id}{A data frame containing appropriate column names}

\item{user_id}{A data frame containing appropriate column names}

\item{access_type}{A data frame containing appropriate column names}
}
\description{
Add signature access information to database
}
