% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkDBTable}
\alias{checkDBTable}
\title{checkDBTable}
\usage{
checkDBTable(conn, db_table_name, check = TRUE)
}
\arguments{
\item{conn}{An established connection to database using SigRepo::newConnhandler()}

\item{db_table_name}{Name of table in the database}

\item{check}{Check whether table exists in the database. Default = TRUE.}
}
\description{
Check api key whether it is valid to access the database
}
\keyword{internal}
