% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkDuplicatedEmails}
\alias{checkDuplicatedEmails}
\title{checkDuplicatedEmails}
\usage{
checkDuplicatedEmails(
  conn,
  db_table_name = "users",
  table,
  coln_var = "user_email",
  check_db_table = TRUE
)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{table}{A data frame object}

\item{coln_var}{A list of column names to be excluded from the check.}

\item{check_db_table}{Check whether table exists in the database. Default = TRUE.}
}
\description{
Check if provided user table has duplicated emails.
}
\keyword{internal}
