% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkOmicSignature}
\alias{checkOmicSignature}
\title{checkOmicSignature}
\usage{
checkOmicSignature(omic_signature, check = TRUE)
}
\arguments{
\item{omic_signature}{An R6 class object from OmicSignature package}
}
\description{
Check omic_signature is a valid R6 object
}
\keyword{internal}
