% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkPermissions}
\alias{checkPermissions}
\title{checkPermissions}
\usage{
checkPermissions(
  conn,
  action_type = c("SELECT", "INSERT", "UPDATE", "DELETE", "CREATE USER"),
  required_role = c("admin", "editor", "viewer")
)
}
\arguments{
\item{conn}{An established connection to database using SigRepo::newConnhandler()}

\item{action_type}{An established connection to database using SigRepo::newConnhandler()}

\item{required_role}{An established connection to database using SigRepo::newConnhandler()}
}
\description{
Check api key whether it is valid to access the database
}
\keyword{internal}
