% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkTableInput}
\alias{checkTableInput}
\title{checkTableInput}
\usage{
checkTableInput(
  conn,
  db_table_name,
  table,
  exclude_coln_names = NULL,
  check_db_table = TRUE
)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{table}{A data frame object}

\item{exclude_coln_names}{A list of column names to be excluded from the check.}

\item{check_db_table}{Check whether table exists in the database. Default = TRUE.}
}
\description{
Check table against a particular table in the database.
}
\keyword{internal}
