% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{clearDropdown}
\alias{clearDropdown}
\title{General function for clearing selections of a dropdown menu}
\usage{
clearDropdown(dropdown, field, dbTable)
}
\arguments{
\item{dropdown}{the id of the dropdown menu to update}

\item{field}{the name of the search field whose dropdown menu should be
cleared}

\item{dbTable}{the database table to query}
}
\description{
General function for clearing selections of a dropdown menu
}
