% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructBetweenClause.R
\name{constructBetweenClause}
\alias{constructBetweenClause}
\title{Assemble part of a where clause of the form
  "<field> BETWEEN <value1> AND <value2>"}
\usage{
constructBetweenClause(myList, listKey)
}
\arguments{
\item{myList}{named list where the names are fields and the values are values
associated with those fields}

\item{listKey}{the field name to access from myList

Note that myList[[listKey]] must have exactly two values}
}
\description{
Assemble part of a where clause of the form
  "<field> BETWEEN <value1> AND <value2>"
}
