% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyFile.R
\name{copyFile}
\alias{copyFile}
\title{copyFile}
\usage{
copyFile(
  objectFile,
  signatureServer = Sys.getenv("signatureServer"),
  uploadPath = Sys.getenv("signatureDirectory"),
  thisUser
)
}
\arguments{
\item{objectFile}{input file object}

\item{signatureServer}{server hosting the signature files. Defaults to
configuration settings}

\item{uploadPath}{the directory to which the signature files get written.
Defaults to configuration settings.}

\item{thisUser}{the username that's uploading the file. Note that this user
needs to be able to write to your target directory.}
}
\description{
copyFile
}
