% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{createHashKey}
\alias{createHashKey}
\title{createHashKey}
\usage{
createHashKey(table, hash_var, hash_columns, hash_method = c("md5", "sodium"))
}
\arguments{
\item{hash_var}{An api key uses to access the database}

\item{hash_columns}{An api key uses to access the database}

\item{hash_method}{An api key uses to access the database}

\item{conn}{An established connection to database using newConnhandler()}
}
\description{
Check api key whether it is valid to access the database
}
\keyword{internal}
