% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOmicCollection.R
\name{createOmicCollection}
\alias{createOmicCollection}
\title{createOmicCollection}
\usage{
createOmicCollection(conn_handler, db_collection_tbl)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database
obtained from SigRepo::newConnhandler() (required)}

\item{db_collection_tbl}{A collection table in the database with associated 
a list of signature ids to be made into an OmicSignatureCollection object (required)}
}
\description{
Get the collection set uploaded by a specific user in the database.
}
\keyword{internal}
