% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOmicSignature.R
\name{createOmicSignature}
\alias{createOmicSignature}
\title{createOmicSignature}
\usage{
createOmicSignature(conn_handler, db_signature_tbl)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database
obtained from SigRepo::newConnhandler() (required)}

\item{db_signature_tbl}{A data frame contains a list of signatures
to be made into an OmicSignature object (required)}
}
\description{
Get the signature set uploaded by a specific user in the database.
}
\keyword{internal}
