% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSignatureMetadata.R
\name{createSignatureMetadata}
\alias{createSignatureMetadata}
\title{createSignatureMetadata}
\usage{
createSignatureMetadata(conn_handler, omic_signature, verbose = FALSE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{omic_signature}{An R6 class object from OmicSignature package}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{FALSE}.}
}
\description{
Create a metadata object for a signature
}
\keyword{internal}
