% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCollection.R
\name{deleteCollection}
\alias{deleteCollection}
\title{deleteCollection}
\usage{
deleteCollection(conn_handler, collection_id, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_id}{ID of the collection being removed from the database (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Delete a collection from the collection table of the database
}
