% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteOrganism.R
\name{deleteOrganism}
\alias{deleteOrganism}
\title{deleteOrganism}
\usage{
deleteOrganism(conn_handler, organism, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{organism}{A list of organisms to be removed (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Remove organisms from database
}
