% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePlatform.R
\name{deletePlatform}
\alias{deletePlatform}
\title{deletePlatform}
\usage{
deletePlatform(conn_handler, platform_name, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{platform_name}{A list of platform names to be removed (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Remove platforms from database
}
