% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSampleType.R
\name{deleteSampleType}
\alias{deleteSampleType}
\title{deleteSampleType}
\usage{
deleteSampleType(conn_handler, sample_type, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{sample_type}{A list of sample types to be removed (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Remove sample types from database
}
