% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignature.R
\name{deleteSignature}
\alias{deleteSignature}
\title{deleteSignature}
\usage{
deleteSignature(signatureName, connHandle, thisUser, force = F)
}
\arguments{
\item{signatureName}{character vector of the name of the target signature}

\item{connHandle}{Database connection handle}

\item{thisUser}{the username of the submitter}

\item{force}{whether to have a confirmation prompt before removing}
}
\description{
Function that executes all required deletess of a signature
}
