% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignatureFromCollection.R
\name{deleteSignatureFromCollection}
\alias{deleteSignatureFromCollection}
\title{deleteSignatureFromCollection}
\usage{
deleteSignatureFromCollection(conn_handler, collection_id, signature_id)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to
a remote database obtained from SigRepo::newConnhandler()}

\item{collection_id}{ID of the collection in the database}

\item{signature_id}{A list of signature IDs to be removed from the 
collection in the databse}
}
\description{
Delete a list of signatures from collection table in the database
}
