% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUser.R
\name{deleteUser}
\alias{deleteUser}
\title{deleteUser}
\usage{
deleteUser(conn_handler, user_name)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{A data frame containing appropriate column names:
user_name, user_password, user_email, user_first, user_last, user_affiliation, user_role}
}
\description{
Delete user information in the database
}
