% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIKey.R
\name{getAPIKey}
\alias{getAPIKey}
\title{getAPIKey}
\usage{
getAPIKey(conn_handler)
}
\arguments{
\item{conn_handler}{An established database connection using newConnhandler()}
}
\description{
Add an API Key for one or more users in the database
}
