% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollection.R
\name{getCollection}
\alias{getCollection}
\title{getCollection}
\usage{
getCollection(
  conn_handler,
  collection_name = NULL,
  collection_id = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_name}{Name of collection to be returned}

\item{collection_id}{ID of collection to be returned}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get a list of collection uploaded by a specified user in the database.
}
\examples{

# Establish a Connection Handler using newConnHandler if not done so already.

# search for available collections

# collection_tbl <- SigRepo::getCollection(
# conn_handler = conn,
# collection_name = "test_collection",
# colection_id = 170,
# verbose = FALSE)

}
