% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{getDBColNames}
\alias{getDBColNames}
\title{getDBColNames}
\usage{
getDBColNames(conn, db_table_name, check_db_table = TRUE)
}
\arguments{
\item{conn}{An established connection to database using SigRepo::newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{check_db_table}{Check whether table exists in the database. Default = TRUE}
}
\description{
Get column names of a particular table in the database
}
\keyword{internal}
