% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures}
\usage{
getFeatures(
  conn,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  filter_by = NULL
)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{filter_by}{author id used to submit the signature}
}
\description{
Get signatures to database
}
