% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{getFieldValues}
\alias{getFieldValues}
\title{General function for getting the values of a field using sqlFindingQuery()}
\usage{
getFieldValues(field, dbTable, ins, betweens)
}
\arguments{
\item{field}{the name of the search field whose dropdown menu should be
updated}

\item{dbTable}{the database table to query}

\item{ins}{list of all of the possible in clauses for sqlFindingQuery()}

\item{betweens}{list of all of the possible between clauses for
sqlFindingQuery()}
}
\description{
General function for getting the values of a field using sqlFindingQuery()
}
