% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2.R
\name{getLevel2}
\alias{getLevel2}
\title{getLevel2}
\usage{
getLevel2(insList = NULL)
}
\arguments{
\item{insList}{a named list of criteria you want to search by,
similar to sqlFindingQuery. defaults to getting every signature}
}
\description{
obtains level2 representation of a signature
}
