% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{getNumOfObs}
\alias{getNumOfObs}
\title{getNumOfObs}
\usage{
getNumOfObs(conn, db_table_name)
}
\arguments{
\item{conn}{An established connection to database using SigRepo::newConnhandler()}

\item{db_table_name}{A table in the database}
}
\description{
Check api key whether it is valid to access the database
}
\keyword{internal}
