% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganisms.R
\name{getOrganisms}
\alias{getOrganisms}
\title{getOrganisms}
\usage{
getOrganisms(conn, filter_by = NULL)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{filter_by}{author id used to submit the signature}
}
\description{
Get signatures to database
}
