% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignatures.R
\name{getSignatures}
\alias{getSignatures}
\title{getSignatures}
\usage{
getSignatures(signatureNames = c("ALL"), withLevel2 = F)
}
\arguments{
\item{signatureNames:}{A character vector with the signature Names. By default,
setting this parameter to "ALL" gets every signature's metadata.}

\item{withLevel2:}{Boolean denoting whether you want to}
}
\value{
a named list of dataframes. "metadata" holds the signatures' metadata.
"level2" holds the level2 representation.
}
\description{
obtains the metadata for a given list of signature Names 
within the database. Note that signature_name is a unique identifier 
in the database. If desired, you can obtain the level2 information for your signatures.
Level2 signature data is simply the list of significant features and their corresponding
weight and direction for a given signature.
}
