% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\title{getSpecies}
\usage{
getSpecies(myIns = NULL)
}
\arguments{
\item{myIns}{optional named list with query parameters. see ?sqlFindingQuery
for more information.}
}
\description{
gets species from the species table in a database
}
