% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableFields.R
\name{getTableFields}
\alias{getTableFields}
\title{getTableFields}
\usage{
getTableFields(tableName)
}
\arguments{
\item{tableName}{character denoting the name of a table in the database.}
}
\description{
obtains the fields of a table of interest within the database.
}
