% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{getVariableID}
\alias{getVariableID}
\title{getVariableID}
\usage{
getVariableID(
  conn,
  db_table_name,
  table,
  coln_var,
  coln_var_id,
  check_db_table = TRUE
)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{db_table_name}{An api key uses to access the database}

\item{table}{An api key uses to access the database}

\item{coln_var}{An api key uses to access the database}

\item{coln_var_id}{An api key uses to access the database}

\item{check_db_table}{An api key uses to access the database}
}
\description{
Get a variable ID by looking up its primary key or hash key
}
\keyword{internal}
