% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivateUser.R
\name{inactivateUser}
\alias{inactivateUser}
\title{inactivateUser}
\usage{
inactivateUser(conn_handler, user_name, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{user_name}{Name of the user to be inactivated (required).}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Inactivate a user in the database
}
