% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFunctions.R
\name{insert_table_sql}
\alias{insert_table_sql}
\title{insert_table_sql}
\usage{
insert_table_sql(conn, db_table_name, table)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{table}{A table in the database}
}
\description{
Insert a table into the database
}
