% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupUser.R
\name{lookupUser}
\alias{lookupUser}
\title{lookupUser}
\usage{
lookupUser(conn, username, password, api_key = TRUE)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{username}{A data frame containing appropriate column names}

\item{password}{A data frame containing appropriate column names}

\item{api_key}{A data frame containing appropriate column names}
}
\description{
lookup a user in the database
}
