% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnHandle.R
\name{newConnHandle}
\alias{newConnHandle}
\title{newConnHandle}
\usage{
newConnHandle(
  db = "sigrepo",
  driver = RMySQL::MySQL(),
  databaseHost = Sys.getenv("databaseServer"),
  databasePort = as.integer(Sys.getenv("databasePort")),
  thisUser = "guest",
  thisPassword = "guest",
  usePassword = "YES"
)
}
\arguments{
\item{db}{character: the schema to which you want the handle to point.}

\item{databaseHost}{character: the server you to which you want to connect}

\item{databasePort}{integer: the Port you wish to use. Our server currently uses
port 4253}

\item{thisUser}{character: the user establishing the handle}

\item{thisPassword}{character: password associated with user. optional and
can be set to null, but
would need to input the password separately through the getPass method defined
in the function.}

\item{usePassword}{character: whether to use the password specified or ask
the user to input the password through getPass.}
}
\value{
a MySQL connection Handle.
}
\description{
establishes a new connection handle to the host
updated 2020/08/12
If you want to use the guest account, has very restricted access, select only.
newConnHandle("You", thisPassword = NULL)
will prompt you to enter password for user "You"
}
