% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectUploadQC.R
\name{objectUploadQC}
\alias{objectUploadQC}
\title{objectUploadQC}
\usage{
objectUploadQC(objectFile)
}
\arguments{
\item{objectFile}{rds or json file representing your OmicSignature Object}
}
\description{
conducts Quality checking of the OmicSignature file to be uploaded
}
