% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddOnesOut.R
\name{oddOnesOut}
\alias{oddOnesOut}
\title{oddOnesOut}
\usage{
oddOnesOut(your_obj)
}
\arguments{
\item{your_obj}{your OmicSignatureObject}
}
\description{
finds features that are not in your database but are in your file
}
