% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeCheckOrAdd.R
\name{phenotypeCheckOrAdd}
\alias{phenotypeCheckOrAdd}
\title{phenotypeCheckOrAdd}
\usage{
phenotypeCheckOrAdd(
  phenotype = NULL,
  thisHandle = newConnHandle(),
  verbose = T
)
}
\arguments{
\item{phenotype}{character vector}

\item{thisHandle}{connection handle to use}

\item{verbose}{Boolean: whether to show more detailed outputs}
}
\description{
checks if phenotype exists in database already and, if not,
inserts into phenotypes table of database
}
