% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{print_messages}
\alias{print_messages}
\title{print_messages}
\usage{
print_messages(verbose)
}
\arguments{
\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Function to whether print diagnostic messages or not
}
\keyword{internal}
