% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protTransform.R
\name{protTransform}
\alias{protTransform}
\title{protTransform}
\usage{
protTransform(
  organism_code = "HUMAN",
  tax_id = "9606",
  version = format(Sys.Date(), "\%m\%d\%Y"),
  is_current = 1
)
}
\arguments{
\item{organism_code}{organism name you want to retrieve data from}

\item{tax_id}{Taxonomic id of organism}

\item{version}{this is set the date of downloading the ftp file}

\item{is_current}{boolean for if the feature_names are current or not.}
}
\description{
Function to dynamically transform the ftp uniprot data into the SigRepo dictionary structure
}
