% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refcheck.R
\name{Refcheck}
\alias{Refcheck}
\title{RefCheck}
\usage{
Refcheck(signature_object, assay)
}
\arguments{
\item{signature_object}{An omic signature object.}

\item{assay}{A string indicating the assay type (e.g., "transcriptomics")}
}
\value{
A data frame of outdated or unresolvable feature_name IDs, if there are any(with optional replacements)
}
\description{
Queries Ensembl API to cross-reference Ensembl feature_name_ids in batches of 1000
}
\examples{
Refcheck(OmS_ex_sig, assay = "transcriptomics")
}
\keyword{internal}
