% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{removeDuplicates}
\usage{
removeDuplicates(conn, db_table_name, table, coln_var, check_db_table = TRUE)
}
\arguments{
\item{conn}{An established connection to database using SigRepo::newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{table}{A data frame object}

\item{coln_var}{A column variable in the data table}

\item{check_db_table}{whether to check database table. Default = TRUE}
}
\description{
Remove duplicates from table in the database
}
\keyword{internal}
