% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSignatureFromCollection.R
\name{removeSignatureFromCollection}
\alias{removeSignatureFromCollection}
\title{removeSignatureFromCollection}
\usage{
removeSignatureFromCollection(conn_handler, collection_id, signature_id)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to
a remote database obtained from SigRepo::newConnhandler()}

\item{collection_id}{ID of collection in the database}

\item{signature_id}{A list of signature IDs to be removed from a collection 
in the database}
}
\description{
Delete a list of signatures from a collection in the database
}
