% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestUser.R
\name{requestUser}
\alias{requestUser}
\title{requestUser}
\usage{
requestUser(newUserName = NULL, newUserEmail = NULL)
}
\arguments{
\item{newUserName}{string denoting your requested user name}

\item{newUserEmail}{string denoting the new user's email}
}
\description{
adds a request for a new user to have write access
to the database that they're pointing to. This function adds a record to the
"user_requests" table, which will then be reviewed by an administrator.
The record will not be added if:
...There is already an existing user under that username
...There is already a request under that username
}
