% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveOmicSigObj.R
\name{retrieveOmicSigObj}
\alias{retrieveOmicSigObj}
\title{retrieve OmicSig obj from VM file system; hard coded}
\usage{
retrieveOmicSigObj(signatureName)
}
\arguments{
\item{signatureName}{}
}
\value{
OmicSignature object
}
\description{
retrieveOmicSigObj()
}
