% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveTranscriptomicsFeatureSet.R
\name{retrieveTranscriptomicsFeatureSet}
\alias{retrieveTranscriptomicsFeatureSet}
\title{retrieveTranscriptomicsFeatureSet}
\usage{
retrieveTranscriptomicsFeatureSet(conn, feature_tbl)
}
\arguments{
\item{conn}{An established connection to the database obtained from SigRepo::conn_init()}

\item{feature_tbl}{A data frame contains a list of feature to look up by}
}
\description{
Search for a list of features based on an assay type in the database
}
\keyword{internal}
