% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCollection.R
\name{searchCollection}
\alias{searchCollection}
\title{searchCollection}
\usage{
searchCollection(
  conn_handler,
  collection_name = NULL,
  collection_id = NULL,
  signature_name = NULL,
  signature_id = NULL,
  user_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_name}{Name of collection to be looked up by.}

\item{collection_id}{ID of collection to be looked up by.}

\item{signature_name}{Name of signatures to be looked up by.}

\item{signature_id}{ID of signatures to be looked up by.}

\item{user_name}{Name of users that the collection belongs to.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get a list of collection available in the database
}
\examples{

# Create a db connection
conn_handler <- SigRepo::newConnHandler(
 dbname = "sigrepo", 
 host = "montilab.bu.edu", 
 port = 3306, 
 user = "guest", 
 password = "guest"
)

# Get a list of collection available in the database
collection_tbl <- sigRepo::searchCollection(
  conn_handler = conn_handler
)

}
