% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFeature.R
\name{searchFeature}
\alias{searchFeature}
\title{searchFeature}
\usage{
searchFeature(
  conn_handler,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  feature_name = NULL
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to a remote database 
obtained from SigRepo::newConnhandler() (required)}

\item{assay_type}{The assay type uses to filter the features by (required). 
Type of assays: transcriptomics, proteomics, metabolomics, methylomics, 
genetic_variations, dna_binding_sites. 
Default: transcriptomics.}

\item{feature_name}{a list of features to search by. Default is NULL which
includes all features in the database.}
}
\description{
Get a list of features available in the database
}
