% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOrganism.R
\name{searchOrganism}
\alias{searchOrganism}
\title{searchOrganism}
\usage{
searchOrganism(conn_handler, organism = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{organism}{a list of organisms to search by. Default is NULL which
includes all organisms in the database.}
}
\description{
Get organisms in the database
}
