% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOrganisms.R
\name{searchOrganisms}
\alias{searchOrganisms}
\title{searchOrganisms}
\usage{
searchOrganisms(conn_handler, organism = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{organism}{a list of organisms to search by. Default is NULL which
will return all of the organisms in the database.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of organisms in the database
}
