% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPhenotype.R
\name{searchPhenotype}
\alias{searchPhenotype}
\title{searchPhenotype}
\usage{
searchPhenotype(conn_handler, phenotype = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{phenotype}{a list of phenotypes to search by. Default is NULL which
includes all phenotypes in the database}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get phenotypes in the database
}
