% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPhenotypes.R
\name{searchPhenotypes}
\alias{searchPhenotypes}
\title{searchPhenotypes}
\usage{
searchPhenotypes(conn_handler, phenotype = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{phenotype}{A list of phenotypes to search by. Default is NULL which
will return all of the phenotypes in the database}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of phenotypes in the database
}
