% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPlatform.R
\name{searchPlatform}
\alias{searchPlatform}
\title{searchPlatform}
\usage{
searchPlatform(conn_handler, platform_id = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{platform_id}{a list of platform accession ids to search by. Default is NULL which
includes all platforms in the database.}
}
\description{
Get signatures to database
}
