% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPlatforms.R
\name{searchPlatforms}
\alias{searchPlatforms}
\title{searchPlatforms}
\usage{
searchPlatforms(conn_handler, platform_name = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{platform_name}{A list of platform names to be looked up. Default is 
NULL which will return all of the platforms in the database.}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of platforms in the database
}
