% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSampleType.R
\name{searchSampleType}
\alias{searchSampleType}
\title{searchSampleType}
\usage{
searchSampleType(conn_handler, sample_type = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{sample_type}{a list of sample types to search by. Default is NULL which
includes all sample types in the database.}
}
\description{
Get sample types in the database
}
