% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSampleTypes.R
\name{searchSampleTypes}
\alias{searchSampleTypes}
\title{searchSampleTypes}
\usage{
searchSampleTypes(
  conn_handler,
  sample_type = NULL,
  brenda_accession = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{sample_type}{A list of sample types to search by. Default is NULL which
will return all of the sample types in the database.}

\item{brenda_accession}{A list of Brenda accession to search by. Default is NULL.}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of sample types in the database
}
