% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTranscriptomicsFeatureSet.R
\name{searchTranscriptomicsFeatureSet}
\alias{searchTranscriptomicsFeatureSet}
\title{searchTranscriptomicsFeatureSet}
\usage{
searchTranscriptomicsFeatureSet(
  conn_handler,
  feature_name = NULL,
  organism = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{feature_name}{A list of feature names to look up.}

\item{organism}{A list of organism to look up.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of transcriptomics features in the database
}
