% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{selectedHTML}
\alias{selectedHTML}
\title{General function for generating html output showing selected search terms}
\usage{
selectedHTML(searchID, displayName)
}
\arguments{
\item{searchID}{the dropdown menu id whose selected values should be
displayed}

\item{displayName}{the word or phrase to display in the html output
representing the dropdown menu field}
}
\description{
General function for generating html output showing selected search terms
}
