% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrorMessages.R
\name{showAssayErrorMessage}
\alias{showAssayErrorMessage}
\title{showAssayTypeErrorMessage}
\usage{
showAssayErrorMessage(assay_type)
}
\arguments{
\item{assay_type}{The unknown assay type}
}
\description{
Error message for trying to add unknown assay types to the database
}
\keyword{internal}
