% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrorMessages.R
\name{showAssayTypeErrorMessage}
\alias{showAssayTypeErrorMessage}
\title{showAssayTypeErrorMessage}
\usage{
showAssayTypeErrorMessage(unknown_values)
}
\arguments{
\item{unknown_values}{The unknown assay type}
}
\description{
Error message for trying to add unknown assay types to the database
}
\keyword{internal}
