% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrorMessages.R
\name{showOrganismErrorMessage}
\alias{showOrganismErrorMessage}
\title{showOrganismErrorMessage}
\usage{
showOrganismErrorMessage(db_table_name, unknown_values)
}
\arguments{
\item{db_table_name}{The table name in database}

\item{unknown_values}{The unknown values}
}
\description{
Error message for trying to add unknown organisms to the database
}
\keyword{internal}
