% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrorMessages.R
\name{showSampleTypeErrorMessage}
\alias{showSampleTypeErrorMessage}
\title{showSampleTypeErrorMessage}
\usage{
showSampleTypeErrorMessage(db_table_name, unknown_values)
}
\arguments{
\item{db_table_name}{The table name in database}

\item{unknown_values}{The unknown values}
}
\description{
Error message for trying to add unknown sample types to the database
}
\keyword{internal}
