% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showErrorMessages.R
\name{showTranscriptomicsErrorMessage}
\alias{showTranscriptomicsErrorMessage}
\title{showTranscriptomicsErrorMessage}
\usage{
showTranscriptomicsErrorMessage(db_table_name, unknown_values)
}
\arguments{
\item{db_table_name}{The table name in database}

\item{unknown_values}{The unknown values}
}
\description{
Error message for trying to add unknown Transcriptomics Features to the database
}
\keyword{internal}
