% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRepoLink.R
\docType{data}
\name{SigRepoLink}
\alias{SigRepoLink}
\title{connection to the VM server link}
\format{
An object of class \code{character} of length 1.
}
\usage{
SigRepoLink
}
\description{
connection to the VM server link
}
\keyword{datasets}
