% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRepoLink.R
\docType{data}
\name{SigRepoSignatureDir}
\alias{SigRepoSignatureDir}
\title{connection to the file system}
\format{
An object of class \code{character} of length 1.
}
\usage{
SigRepoSignatureDir
}
\description{
connection to the file system
}
\keyword{datasets}
