% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singeQuote.R
\name{singleQuote}
\alias{singleQuote}
\title{Add single quotation marks around a string}
\usage{
singleQuote(myString)
}
\arguments{
\item{myString}{the string to add single quotation marks to}
}
\description{
Add single quotation marks around a string
}
