% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFindingQuery.R
\name{sqlFindingQuery}
\alias{sqlFindingQuery}
\title{Constructs sql query based on where clause, if one is needed,
  and executes final query as output}
\usage{
sqlFindingQuery(
  dbTable,
  fields = c("*"),
  ins = NULL,
  betweens = NULL,
  orderFields = NULL,
  distinct = TRUE,
  verbose = FALSE,
  collapse_by = " AND ",
  handle = newConnHandle()
)
}
\arguments{
\item{dbTable}{string; the table to select from}

\item{fields}{character vector, the fields to select from the target table}

\item{ins}{named list where the names are the fields to narrow search by
and the values are vectors of the values to look for in those fields}

\item{betweens}{named list where the names are the fields to narrow search by
and the values are vectors of length 2 containing the endpoints of the
range to consider in those fields}

\item{orderFields}{a vector of fields to order the query results by}

\item{distinct}{a boolean indicating whether to return unique results}
}
\description{
Constructs sql query based on where clause, if one is needed,
  and executes final query as output
}
