% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlGeneric.R
\name{sqlGeneric}
\alias{sqlGeneric}
\title{Submit a query to the database}
\usage{
sqlGeneric(query, conn = newConnHandle(), disconnectAfter = T)
}
\arguments{
\item{query}{the query to submit}
}
\description{
Submit a query to the database
}
