% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsByGroup.R
\name{statsByAttribute}
\alias{statsByAttribute}
\title{statsByAttribute}
\usage{
statsByAttribute(attribute)
}
\arguments{
\item{attribute}{the field name of which you want to get signature counts.
Keep in mind the possible attributes depend on the fields declared in your
instance of SigRepoR, if you have one. Possible legal entries include
'phenotype', 'organism', 'assay_platform'. uses platform_signature_view
as reference view.}
}
\value{
Dataframe with the following fields:
> "attribute"-column containing each possible value of said attribute
> "count"-the number of signatures belonging to each attribute
}
\description{
obtain counts of signatures of each possible attribute of
interest in the Database
updated 2020/08/12
}
