% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSigCompare.R
\name{twoSigCompare}
\alias{twoSigCompare}
\title{Compare two lists of signatures}
\usage{
twoSigCompare(
  sig1,
  sig2,
  sig1_name = "sig1",
  sig2_name = "sig2",
  background_number = 22000
)
}
\arguments{
\item{sig1}{a character vector for signature 1}

\item{sig2}{a character vector for signature 2}

\item{sig1_name}{string. the name of signature 1. used for display in Venn diagram.}

\item{sig2_name}{string. the name of signature 2. used for display in Venn diagram.}

\item{background_number}{a numeric value indicating the total number of features. For example, in microarray, there are usually 22000 genes on a chip.}
}
\value{
a list of the comparison result of the two signature.
including: "sig1_name", "sig2_name", "sig1_symbol", "sig2_symbol",
"only_sig1", "only_sig2", "sig_both", "Venn", "hyper_p.value"
}
\description{
Compare two lists of signatures and return Venn diagram and hypergeometric test p-value for intersect.
updated 06/2020
}
\examples{
sig1 <- c("TNMD", "DPM1", "CFH", "FUCA2", "HECW1", "LASP1", "MAD1L1", "KLHL13", "FUCA2")
sig2 <- c("TNMD", "FUCA2", "C1orf112", "FGR", "MAD1L1", "NFYA", "STPG1", "CFTR", "LAS1L", "ENPP4", "SEMA3F", "ANKIB1")
compare_result <- twoSigCompare(sig1, sig2, sig1_name = "sig1", sig2_name = "sig2", is.lv2 = FALSE, background_number = 22000)
compare_result$Venn
compare_result$sig_both
}
