% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSigKS.R
\name{twoSigKSEnrichment}
\alias{twoSigKSEnrichment}
\title{Enrichment test via one-sided Kolmogorov–Smirnov test of two signatures}
\usage{
twoSigKSEnrichment(
  signature,
  signatureGset,
  weights = NULL,
  weights.pwr = 1,
  absolute = FALSE,
  plotting = TRUE
)
}
\arguments{
\item{signature}{A vector of ranked symbols, or a signature dataframe}

\item{signatureGset}{another signature, vector or signature dataframe,
as the "gene set" to be compared to}

\item{weights}{Weights for weighted score, not needed if input is signature dataframe}

\item{weights.pwr}{Exponent for weights}

\item{absolute}{Takes max-min score rather than the max deviation from null}

\item{plotting}{Use true to generate plot}
}
\value{
A list of data and plots
}
\description{
Enrichment test via one-sided Kolmogorov–Smirnov test of two signatures
}
