% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCollectionEntity.R
\name{updateCollectionEntity}
\alias{updateCollectionEntity}
\title{updateCollectionEntity}
\usage{
updateCollectionEntity(
  connHandle,
  collectionName = NULL,
  newCollectionName = NULL,
  newCollectionDescription = NULL
)
}
\arguments{
\item{connHandle}{DB connection}

\item{collectionName}{current collection name}

\item{newCollectionName}{new collection name}

\item{newCollectionDescription}{new collection description}
}
\value{
Dataframe with the following fields:
> "attribute"-column containing each possible value of said attribute
> "count"-the number of signatures belonging to each attribute
}
\description{
update a collection in the DB. ensure you have update privilages
}
