% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{updateDropdown}
\alias{updateDropdown}
\title{General function for updating a dropdown menu}
\usage{
updateDropdown(dropdown, field, dbTable, ins, betweens)
}
\arguments{
\item{dropdown}{the id of the dropdown menu to update}

\item{field}{the name of the search field whose dropdown menu should be
updated}

\item{dbTable}{the database table to query}

\item{ins}{list of all of the possible in clauses for sqlFindingQuery()}

\item{betweens}{list of all of the possible between clauses for
sqlFindingQuery()}
}
\description{
General function for updating a dropdown menu
}
