% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSignature.R
\name{updateSignature}
\alias{updateSignature}
\title{updateSignature}
\usage{
updateSignature(conn, omic_signature)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{omic_signature}{An R6 class object from OmicSignature package}

\item{signature_name}{name of your signature that you previously uploaded}
}
\description{
Add a signature to signature table of the database
}
