% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUser.R
\name{updateUser}
\alias{updateUser}
\title{updateUser}
\usage{
updateUser(
  conn_handler,
  user_name,
  user_email = NULL,
  user_affiliation = NULL,
  user_first = NULL,
  user_last = NULL,
  user_role = c("admin", "editor", "viewer")
)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{A data frame containing appropriate column names:
user_name, user_password, user_email, user_first, user_last, user_affiliation, user_role}
}
\description{
Update user information in the database
}
