% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateUser.R
\name{validateUser}
\alias{validateUser}
\title{validateUser}
\usage{
validateUser(conn_handler)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to a remote database 
obtained from SigRepo::newConnhandler() (required)}
}
\description{
Validate a user in the database
}
