% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSignatureFile.R
\name{writeSignatureFile}
\alias{writeSignatureFile}
\title{writeSignatureFile}
\usage{
writeSignatureFile(fileObject, userName)
}
\arguments{
\item{fileObject}{signature object}

\item{userName}{name of the user uploading}
}
\description{
writeSignatureFile
}
