% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCollection.R
\name{addCollection}
\alias{addCollection}
\title{addCollection}
\usage{
addCollection(
  conn_handler,
  omic_collection,
  visibility = FALSE,
  return_collection_id = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{omic_collection}{A collection of OmicSignature objects from OmicSignature package}

\item{visibility}{A logical value indicates whether or not to allow others 
to view and access one's uploaded collection. Default is \code{FALSE}.}

\item{return_collection_id}{a logical value indicates whether or not to return
the ID of the uploaded collection. Default is \code{FALSE}.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add signature collection to database
}
