% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKeyword.R
\name{addKeyword}
\alias{addKeyword}
\title{addKeyword}
\usage{
addKeyword(conn_handler, keyword_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{keyword_tbl}{A data frame containing appropriate column names: 
keyword (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add keywords to database
}
