% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKeyword.R
\name{addKeyword}
\alias{addKeyword}
\title{addKeyword}
\usage{
addKeyword(conn, keyword_tbl)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{keyword_tbl}{A data frame containing appropriate column names: keyword}
}
\description{
Add keywords to database
}
