% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMoreInformation.R
\name{addMoreInformation}
\alias{addMoreInformation}
\title{addMoreInformation}
\usage{
addMoreInformation(
  signatureName,
  connHandle = newConnHandle(),
  rootDir = Sys.getenv("signatureDirectory"),
  objectExtension = "_obj.json",
  verbose = FALSE
)
}
\arguments{
\item{signatureName}{name of the signature}

\item{connHandle}{connection to the DB}

\item{rootDir}{directory of the signatures}

\item{verbose}{default to FALSE. use TRUE to print messages}
}
\description{
adds a JSON typed value into the DB associated with
a signature of interest
}
