% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOrganism.R
\name{addOrganism}
\alias{addOrganism}
\title{addOrganism}
\usage{
addOrganism(conn_handler, organism_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{organism_tbl}{A data frame object contains appropriate column names: organism}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add organism to database
}
\examples{


# Creating an dataframe with the approrpiate columns
# organism_tbl <- data.frame(
# organism = "test_organism")

# SigRepo:: addOrganism(
# conn_hanler = conn,
# organism_tbl = organism_tbl,
# verbose = FALSE)

}
