% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOrganism.R
\name{addOrganism}
\alias{addOrganism}
\title{addOrganism}
\usage{
addOrganism(conn_handler, organism_tbl)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{organism_tbl}{A data frame object contains appropriate column names: organism}
}
\description{
Add organism to database
}
