% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhenotype.R
\name{addPhenotype}
\alias{addPhenotype}
\title{addPhenotype}
\usage{
addPhenotype(conn_handler, phenotype_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{phenotype_tbl}{An data frame containing appropriate column names: phenotype}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add phenotype to database
}
