% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlatform.R
\name{addPlatform}
\alias{addPlatform}
\title{addPlatform}
\usage{
addPlatform(conn, platform_tbl)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{platform_tbl}{An data frame containing appropriate column names: 
platform_id, platform_name, seq_technology, organisms to be uploaded into 
the database.}
}
\description{
Add platform to database
}
