% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProteomicsSignatureSet.R
\name{addProteomicsSignatureSet}
\alias{addProteomicsSignatureSet}
\title{addProteomicsSignatureSet}
\usage{
addProteomicsSignatureSet(
  conn_handler,
  signature_id,
  organism_id,
  signature_set,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{signature_id}{ID of the signature (required)}

\item{organism_id}{ID of the organism (required)}

\item{signature_set}{A data frame containing the appropriate column names:
feature_name, probe_id, score, group_label (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add proteomics signature feature set to database
}
\keyword{internal}
