% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleType.R
\name{addSampleType}
\alias{addSampleType}
\title{addSampleType}
\usage{
addSampleType(conn_handler, sample_type_tbl, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{sample_type_tbl}{An data frame containing appropriate column names:
sample_type, brenda_accession (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add sample types to database
}
