% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{addSignature}
\alias{addSignature}
\title{addSignature}
\usage{
addSignature(
  sig_name,
  species_name,
  platform_name,
  cell_line,
  phenotype = NULL,
  uploader,
  uploadHandle,
  verbose = F,
  disconnectAfter = T
)
}
\arguments{
\item{sig_name}{name of signature}

\item{species_name}{name of *species*}

\item{platform_name}{name of assay platform}

\item{cell_line}{name of cell line}

\item{phenotype}{character describing the phenotype of signature}

\item{uploader}{character: the user uploading the signature}

\item{uploadHandle}{connection handle: the connection handle to use
for uploading the signature}

\item{verbose}{default to FALSE. use TRUE to print messages}

\item{disconnectAfter}{Boolean: whether to disconnect the handle
when finished. 
all fields are usually obtained from input values from application}
}
\description{
Adds signature information to signatures table in the DB
}
