% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{addSignature}
\alias{addSignature}
\title{addSignature}
\usage{
addSignature(conn_handler, omic_signature)
}
\arguments{
\item{conn_handler}{An established connection to database using newConnhandler()}

\item{omic_signature}{An R6 class object from OmicSignature package}
}
\description{
Add signature to database
}
