% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{addSignature}
\alias{addSignature}
\title{addSignature}
\usage{
addSignature(
  conn_handler,
  omic_signature,
  visibility = FALSE,
  return_signature_id = FALSE,
  return_missing_features = FALSE,
  verbose = TRUE,
  add_users = NULL
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{omic_signature}{An R6 class object from OmicSignature package (required)}

\item{visibility}{A logical value indicates whether or not to allow others
to view and access one's uploaded signature. Default is \code{FALSE}.}

\item{return_signature_id}{A logical value indicates whether or not to return
the ID of the uploaded signature. Default is \code{FALSE}.}

\item{return_missing_features}{A logical value indicates whether or not to return
a list of features that does not exist in the database. Default is \code{FALSE}.}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}

\item{add_users}{if visibility is set to FALSE, then the user can use this parameter to add a dataframe of users and their corresponding access types to the signature.}
}
\description{
Add signature to database
}
