% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureCollection.R
\name{addSignatureCollection}
\alias{addSignatureCollection}
\title{addSignatureCollection}
\usage{
addSignatureCollection(
  OmicSignatureCollectionObj,
  connHandle,
  uploadPath,
  thisUser
)
}
\arguments{
\item{OmicSignatureCollectionObj}{the OmicSignatureCollection Object to upload}

\item{connHandle}{The database connection handle}

\item{uploadPath}{where to upload the file}

\item{thisUser}{username of the submitter, for confirmation}
}
\description{
adding an OmicSignatureCollection objecct to your file system
and databasee
}
