% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureCollection.R
\name{addSignatureCollection}
\alias{addSignatureCollection}
\title{addSignatureCollection}
\usage{
addSignatureCollection(conn, omic_signature_collection, user_id)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{omic_signature_collection}{A collection of OmicSignature objects from OmicSignature package}

\item{user_id}{An user id}
}
\description{
Add signature collection to database
}
