% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureFeatureSet.R
\name{addSignatureFeatureSet}
\alias{addSignatureFeatureSet}
\title{addSignatureFeatureSet}
\usage{
addSignatureFeatureSet(
  conn,
  signature_id,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  sig_feature_set
)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{signature_id}{An unique signature name}

\item{assay_type}{Type of assays: transcriptomics, proteomics, metabolomics, 
methylomics, genetic_variations, dna_binding_sites}

\item{sig_feature_set}{A data frame containing appropriate column names:
feature_name, origin_feature_id, score, direction}
}
\description{
Add signature feature set to database
}
