% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignatureToCollection.R
\name{addSignatureToCollection}
\alias{addSignatureToCollection}
\title{addSignatureToCollection}
\usage{
addSignatureToCollection(conn_handler, collection_id, signature_id)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to
a remote database obtained from SigRepo::newConnhandler()}

\item{collection_id}{ID of collection in the database}

\item{signature_id}{A list of signature IDs to be added to a collection 
in the database}
}
\description{
Add a list of signatures to a collection in the database
}
