% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscriptomicsFeatureSet.R
\name{addTranscriptomicsFeatureSet}
\alias{addTranscriptomicsFeatureSet}
\title{addTranscriptomicsFeatureSet}
\usage{
addTranscriptomicsFeatureSet(conn, feature_set)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{feature_set}{A data frame containing appropriate column names: 
feature_name, organism, description, synonyms, n_synonyms, ensemble_ids, 
n_ensemble_ids, transcript_biotypes, chromosome_name, start_position, 
end_position}
}
\description{
Add Transcriptomics Feature Set into database
}
