% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscriptomicsFeatureSet.R
\name{addTranscriptomicsFeatureSet}
\alias{addTranscriptomicsFeatureSet}
\title{addTranscriptomicsFeatureSet}
\usage{
addTranscriptomicsFeatureSet(conn_handler, feature_set, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{feature_set}{A data frame containing appropriate column names: 
feature_name, organism, gene_symbol, is_current, version (required)}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add Transcriptomics Feature Set to database
}
\keyword{internal}
