% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUser.R
\name{addUser}
\alias{addUser}
\title{addUser}
\usage{
addUser(conn, user_tbl)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{user_tbl}{A data frame containing appropriate column names:
user_id, user_password, user_email, user_first, user_last, user_affiliation, user_role}
}
\description{
Add user information to database
}
