% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUser.R
\name{addUser}
\alias{addUser}
\title{addUser}
\usage{
addUser(conn_handler, user_tbl, active = FALSE, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{user_tbl}{A data frame containing appropriate column names: user_name, 
user_password, user_email, user_first, user_last, user_affiliation, user_role}

\item{active}{Whether to make a user TRUE (active) or FALSE (inactive). 
Default is \code{FALSE}.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add a list of users to database
}
\examples{

 # creating a user dataframe

# user_tbl <- data.frame(
# user_name = "new_user",
# user_password = "password123",
# user_email = "test_email123",
# user_first = "First",
# user_last = "Last",
# user_affiliation = "University",
# user_role = "editor")


# SigRepo::addUser(
# conn_handler = conn,
# user_tbl = user_tbl,
# verbose = FALSE
# )

}
