% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUser.R
\name{addUser}
\alias{addUser}
\title{addUser}
\usage{
addUser(
  newUserName = NULL,
  newUserEmail = NULL,
  newUserPassword = "NULL",
  connectionHandle
)
}
\arguments{
\item{newUserName}{A character vector with the new username.}

\item{newUserEmail}{A character vector with the new user's email}

\item{newUserPassword}{A character vector with the new user's password.
Strongly discouraged, will default to input prompts}

\item{connectionHandle}{your connection handle object. be sure the user
has "admin" privileges.}
}
\value{
nothing.
}
\description{
adds a user to your SigRepo system
}
