% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUserToCollection.R
\name{addUserToCollection}
\alias{addUserToCollection}
\title{addUserToCollection}
\usage{
addUserToCollection(
  conn_handler,
  collection_id,
  user_name,
  access_type = c("owner", "editor", "viewer"),
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_id}{An ID of collection in the database (required)}

\item{user_name}{A list of users to be added to a collection (required)}

\item{access_type}{A list of permissions to be given to users in order for 
them to view or manage the collection in the database (required) . 

There are three types of permissions:

\code{owner} has Read and Write access to their own uploaded collection.
\code{editor} has Read and Write access to collection that other users were 
given them access to.
\code{viewer} has ONLY Read access to collection that other users were 
given them access to.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Add a list of users with specific access to a collection 
in the database
}
