% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUserToSignature.R
\name{addUserToSignature}
\alias{addUserToSignature}
\title{addUserToSignature}
\usage{
addUserToSignature(
  conn,
  signature_id,
  user_id,
  access_type = c("owner", "user")
)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{signature_id}{signature id}

\item{user_id}{user id}

\item{access_type}{access type: 'owner' or 'user'}
}
\description{
Add user to signature access table in database
}
