% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureSigRepo.R
\name{configureSigRepo}
\alias{configureSigRepo}
\title{configureSigRepo}
\usage{
configureSigRepo(signatureDirectory, databaseServer, databasePort)
}
\arguments{
\item{signatureDirectory}{where your signature files will be}

\item{databaseServer}{where your database is}

\item{databasePort}{which port to use to connect to your database}
}
\description{
configures your instance of Sigrepo to have all its
functions point to the server of your choice by setting System level 
variable assignments
}
