% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructInClause.R
\name{constructInClause}
\alias{constructInClause}
\title{Assemble part of a where clause of the form "<field> IN (<field_values>)"}
\usage{
constructInClause(myList, listKey)
}
\arguments{
\item{myList}{named list where the names are fields and the values are values
associated with those fields}

\item{listKey}{the field name to access from myList}
}
\description{
Assemble part of a where clause of the form "<field> IN (<field_values>)"
}
