% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCollectionMetadata.R
\name{createCollectionMetadata}
\alias{createCollectionMetadata}
\title{createCollectionMetadata}
\usage{
createCollectionMetadata(conn_handler, omic_collection)
}
\arguments{
\item{conn_handler}{An established connection to database using SigRepo::conn_init()}

\item{omic_collection}{An R6 class object from OmicSignature package}
}
\description{
Create a metadata object for a signature
}
\keyword{internal}
