% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{createHashKey}
\alias{createHashKey}
\title{createHashKey}
\usage{
createHashKey(table, hash_var, hash_columns, hash_method = "md5")
}
\arguments{
\item{table}{A data frame with specific columns to be used to create 
the hash key}

\item{hash_var}{The hash variable to be created}

\item{hash_columns}{The specific columns to be used to create the hash key}

\item{hash_method}{The hashing method to create the key. Default \code{md5}.}
}
\description{
Check api key whether it is valid to access the database
}
\keyword{internal}
