% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLink.R
\name{createLink}
\alias{createLink}
\title{Create a link to a signature file based on the signature name}
\usage{
createLink(signatureName)
}
\arguments{
\item{signatureName}{the name of the signature to link to}
}
\description{
Create a link to a signature file based on the signature name
}
