% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFunctions.R
\name{delete_table_sql}
\alias{delete_table_sql}
\title{delete_table_sql}
\usage{
delete_table_sql(
  conn,
  db_table_name,
  delete_coln_var,
  delete_coln_val,
  check_db_table = TRUE
)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{delete_coln_var}{A column variable in the table for removing rows}

\item{delete_coln_val}{A list of values associated with delete_coln_var to be removed.}

\item{check_db_table}{whether to check database table. Default = TRUE.}
}
\description{
delete an entry from database table
}
\keyword{internal}
