% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCollection.R
\name{deleteCollection}
\alias{deleteCollection}
\title{deleteCollection}
\usage{
deleteCollection(conn_handler, collection_id)
}
\arguments{
\item{conn_handler}{An established connection to the database using newConnhandler()}

\item{collection_id}{The name of the signature being deleted from the database}
}
\description{
Delete a signature from the signature table of the database
}
