% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignature.R
\name{deleteSignature}
\alias{deleteSignature}
\title{deleteSignature}
\usage{
deleteSignature(conn_handler, signature_id, verbose = TRUE)
}
\arguments{
\item{conn_handler}{An established connection to the database using newConnhandler()}

\item{signature_id}{ID of signature to be removed from the database}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Delete a signature from the signature table of the database
}
\examples{
# Establish a Connection Handler using newConnHandler if not done so already .


# Get a list of signatures available in the database
# signature_tbl <- SigRepo::searchSignature(
 #  conn_handler = conn,
 #  user_name = "guest"
# )

# Delete signature from database (NOT RUN)
# SigRepo::deleteSignature(
#   conn_handler = conn,
#   collection_id = signature_tbl$signature_id[1]
# )
}
