% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignature.R
\name{deleteSignature}
\alias{deleteSignature}
\title{deleteSignature}
\usage{
deleteSignature(conn_handler, signature_id)
}
\arguments{
\item{conn_handler}{An established connection to the database using newConnhandler()}

\item{signature_id}{ID of signature to be removed from the database}
}
\description{
Delete a signature from the signature table of the database
}
