% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignature.R
\name{deleteSignature}
\alias{deleteSignature}
\title{deleteSignature}
\usage{
deleteSignature(conn, omic_signature)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{signature_name}{name of your signature that you previously uploaded}
}
\description{
Add a signature to signature table of the database
}
