% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUser.R
\name{deleteUser}
\alias{deleteUser}
\title{deleteUser}
\usage{
deleteUser(conn_handler, user_name)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{Name of a user to be deleted (required).}
}
\description{
Delete user information in the database
}
