% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUser.R
\name{deleteUser}
\alias{deleteUser}
\title{deleteUser}
\usage{
deleteUser(conn_handler, user_name, verbose = TRUE)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{Name of a user to be deleted (required).}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Delete user information in the database
}
\examples{


# Establish a Connection Handler using newConnHnalder if not done so already.

# SigRepo::deleteUser(
# conn_handler = conn,
# user_name = "user1",
# verbose = FALSE)

}
