% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUser.R
\name{deleteUser}
\alias{deleteUser}
\title{deleteUser}
\usage{
deleteUser(conn_handler, user_name, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{user_name}{Name of the user to be deleted (required).}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Delete a user from the database
}
