% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIKey.R
\name{getAPIKey}
\alias{getAPIKey}
\title{getAPIKey}
\usage{
getAPIKey(conn_handler)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}
}
\description{
Get API Key of a specific user in the database
}
