% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollection.R
\name{getCollection}
\alias{getCollection}
\title{getCollection}
\usage{
getCollection(conn_handler, collection_name = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_name}{The name of collection to be returned}
}
\description{
Get a list of collection uploaded by a specified user in the database.
}
