% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures}
\usage{
getFeatures(nicely = F)
}
\arguments{
\item{nicely}{boolean: whether to query a feature view that looks nicer.}
}
\description{
obtains the assay platforms within the database.
}
