% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignature.R
\name{getSignature}
\alias{getSignature}
\title{getSignature}
\usage{
getSignature(conn_handler, signature_name = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{signature_name}{The name of signatures to be returned}
}
\description{
Get a list of signatures uploaded by a specified user in the database.
}
