% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignature.R
\name{getSignature}
\alias{getSignature}
\title{getSignature}
\usage{
getSignature(
  conn_handler,
  signature_name = NULL,
  signature_id = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{signature_name}{Name of signatures to be returned}

\item{signature_id}{ID of signatures to be returned}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get a list of signatures uploaded by a specified user in the database.
}
