% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollection.R, R/getSignature.R
\name{getSignature}
\alias{getSignature}
\title{getSignature}
\usage{
getSignature(conn_handler, signature_name = NULL)

getSignature(conn_handler, signature_name = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{signature_name}{The name of a signature that belongs to a specific user 
(\code{user_name}) who previously uploaded the signature into the database.}

\item{user_name}{The name of the user who uploaded the signature into the database.}
}
\description{
Get a list of signatures uploaded by a specified user in the database.

Get a list of signatures uploaded by a specified user in the database.
}
