% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignature.R
\name{getSignature}
\alias{getSignature}
\title{getSignature}
\usage{
getSignature(conn, return_var)
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{author_id}{An author id used to submit the signature}
}
\description{
Get signatures to database
}
