% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignatures.R
\name{getSignatures}
\alias{getSignatures}
\title{getSignatures}
\usage{
getSignatures(conn, author_id = "all")
}
\arguments{
\item{conn}{An established connection to database using newConnhandler()}

\item{author_id}{An author id used to submit the signature}
}
\description{
Get signatures to database
}
