% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFunctions.R
\name{insert_table_sql}
\alias{insert_table_sql}
\title{insert_table_sql}
\usage{
insert_table_sql(conn, db_table_name, table, check_db_table = TRUE)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{db_table_name}{Name of a table in the database}

\item{table}{A table in the database}

\item{check_db_table}{whether to check database table. Default = TRUE}
}
\description{
Insert a table into the database
}
