% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnHandler.R
\name{newConnHandler}
\alias{newConnHandler}
\title{newConnHandler}
\usage{
newConnHandler(
  dbname = "sigrepo",
  host = "montilab.bu.edu",
  port = 3306,
  user = "guest",
  password = "guest"
)
}
\arguments{
\item{dbname}{table schema of which you want the handle to point.}

\item{host}{host server in which you want to connect}

\item{port}{host port in which you wish to use}

\item{user}{the user who is establishing the connection}

\item{password}{password associated with the user}

\item{driver}{driver of the database. Default is RMySQL::MySQL()}
}
\value{
a list with user credentials to establish connection to a remote database
}
\description{
create a handler to connect to a remote database
}
