% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnHandler.R
\name{newConnHandler}
\alias{newConnHandler}
\title{newConnHandler}
\usage{
newConnHandler(
  driver = RMySQL::MySQL(),
  dbname = Sys.getenv("DBNAME"),
  host = Sys.getenv("HOST"),
  port = as.integer(Sys.getenv("PORT")),
  user = Sys.getenv("USER"),
  password = Sys.getenv("PASSWORD")
)
}
\arguments{
\item{driver}{driver of the database. Default is RMySQL::MySQL()}

\item{dbname}{table schema of which you want the handle to point.}

\item{host}{host server in which you want to connect}

\item{port}{host port in which you wish to use}

\item{user}{the user who is establishing the connection}

\item{password}{password associated with the user}
}
\value{
a MySQL connection Handle.
}
\description{
Establish a new connection handle to the host
}
