% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnHandler.R
\name{newConnHandler}
\alias{newConnHandler}
\title{newConnHandler}
\usage{
newConnHandler(
  dbname = "sigrepo",
  host = "montilab.bu.edu",
  port = 3306,
  api_port = 8020,
  user = "guest",
  password = "guest"
)
}
\arguments{
\item{dbname}{Name of MySQL database to point to.}

\item{host}{Name of the server where MySQL database is hosted on.}

\item{port}{Port on the server to connect to MySQL database.}

\item{api_port}{Port on the server to access database API.}

\item{user}{Name of user to establish the connection.}

\item{password}{Password associated with the user.}
}
\value{
A list of user credentials to establish connection to the remote database.
}
\description{
Create a handler to connect to a remote database.
}
\examples{

# the easiest way to do this is use sys.getenv if you have the environment variables set correctly

# conn <- SigRepo::newConnHandler(
# dbname = "sigrepo",
# host = Sys.getenv("HOST"),  
# port = as.integer(Sys.getenv("PORT")), 
# user = Sys.getenv("USER"),
# password = Sys.getenv("PASSWORD") 
# )



}
