% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSignature.R
\name{OmicSignature}
\alias{OmicSignature}
\title{OmicSignature R6 object}
\description{
a R6 object to store signatures generated from experiments. In cluding metadata, signature, and an optional differential expression matrix.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-verbose}{\code{OmicSignature$verbose()}}
\item \href{#method-new}{\code{OmicSignature$new()}}
\item \href{#method-print}{\code{OmicSignature$print()}}
\item \href{#method-extract.signature}{\code{OmicSignature$extract.signature()}}
\item \href{#method-checkMetadata}{\code{OmicSignature$checkMetadata()}}
\item \href{#method-checkSignature}{\code{OmicSignature$checkSignature()}}
\item \href{#method-checkDifexp}{\code{OmicSignature$checkDifexp()}}
\item \href{#method-clone}{\code{OmicSignature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verbose"></a>}}
\subsection{Method \code{verbose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$verbose(v, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$new(metadata, signature, difexp = NULL, print_message = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract.signature"></a>}}
\subsection{Method \code{extract.signature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$extract.signature(conditions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMetadata"></a>}}
\subsection{Method \code{checkMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$checkMetadata(metadata, v = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkSignature"></a>}}
\subsection{Method \code{checkSignature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$checkSignature(
  omicObj,
  signatureType = NULL,
  category_num = 0,
  v = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkDifexp"></a>}}
\subsection{Method \code{checkDifexp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$checkDifexp(difexp, v = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
