% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSignature.R
\name{OmicSignatureCollection}
\alias{OmicSignatureCollection}
\title{OmicSignatureCollection R6 object}
\description{
a R6 object to store a collection of OmicSignature objects.
In cluding metadata, OmicSigList which is a list of OmicSignature object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-verbose}{\code{OmicSignatureCollection$verbose()}}
\item \href{#method-new}{\code{OmicSignatureCollection$new()}}
\item \href{#method-print}{\code{OmicSignatureCollection$print()}}
\item \href{#method-extract.signature}{\code{OmicSignatureCollection$extract.signature()}}
\item \href{#method-checkCollectionMetadata}{\code{OmicSignatureCollection$checkCollectionMetadata()}}
\item \href{#method-checkCollectionOmicSigList}{\code{OmicSignatureCollection$checkCollectionOmicSigList()}}
\item \href{#method-clone}{\code{OmicSignatureCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verbose"></a>}}
\subsection{Method \code{verbose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$verbose(v, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$new(metadata, OmicSigList, print_message = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract.signature"></a>}}
\subsection{Method \code{extract.signature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$extract.signature(conditions, bind = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkCollectionMetadata"></a>}}
\subsection{Method \code{checkCollectionMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$checkCollectionMetadata(metadata, v = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkCollectionOmicSigList"></a>}}
\subsection{Method \code{checkCollectionOmicSigList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$checkCollectionOmicSigList(OmicSigList, v = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignatureCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
