% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveOmicSigObj.R
\name{retrieveOmicSigObj}
\alias{retrieveOmicSigObj}
\title{retrieve OmicSig obj from VM file system}
\usage{
retrieveOmicSigObj(signatureName, signatureDir = SigRepoSignatureDir)
}
\arguments{
\item{signatureName}{name of the signature}

\item{signatureDir}{directory which the signature file is in}
}
\value{
OmicSignature object
}
\description{
retrieve OmicSig obj from VM file system
}
