% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCollection.R
\name{searchCollection}
\alias{searchCollection}
\title{searchCollection}
\usage{
searchCollection(
  conn_handler,
  collection_name = NULL,
  collection_id = NULL,
  signature_name = NULL,
  signature_id = NULL,
  user_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{collection_name}{Name of collection to be looked up by.}

\item{collection_id}{ID of collection to be looked up by.}

\item{signature_name}{Name of the signatures to be looked up by.}

\item{signature_id}{ID of the signatures to be looked up by.}

\item{user_name}{Name of users that the collection belongs to.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get a list of collection available in the database
}
\examples{

# Establish a Connection Handler using newConnHandler if not done so already.

 # SigRepo::searchCollection(
# conn_handler = conn,
# collection_name = "test_collection",
# collection_id = 101,
# signature_name = "test_signature",
# signature_id = 170,
# user_name = "test_user",
# verbose = FALSE
# )



}
