% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFeature.R
\name{searchFeature}
\alias{searchFeature}
\title{searchFeature}
\usage{
searchFeature(
  conn_handler,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  feature_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{assay_type}{Type of assays: transcriptomics, proteomics, metabolomics, 
methylomics, genetic_variations, dna_binding_sites (required)}

\item{feature_name}{A list of feature names to look up.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Search for a list of features based on an assay type in the database
}
