% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPhenotype.R
\name{searchPhenotype}
\alias{searchPhenotype}
\title{searchPhenotype}
\usage{
searchPhenotype(conn_handler, phenotype = NULL)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{phenotypes}{a list of phenotypes to search by. Default is NULL which
includes all phenotypes in the database}
}
\description{
Get phenotypes in the database
}
