% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPlatform.R
\name{searchPlatform}
\alias{searchPlatform}
\title{searchPlatform}
\usage{
searchPlatform(conn_handler, platform_id = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{platform_id}{a list of platfrom ids (platform names) in the database}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get signatures to database
}
