% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSignature.R
\name{searchSignature}
\alias{searchSignature}
\title{searchSignature}
\usage{
searchSignature(
  conn_handler,
  assay_type = c("transcriptomics", "proteomics", "metabolomics", "methylomics",
    "genetic_variations", "dna_binding_sites"),
  user_name = NULL,
  signature_name = NULL,
  organism = NULL,
  phenotype = NULL,
  sample_type = NULL
)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{assay_type}{The assay type uses to filter the features by (required). 
Type of assays: transcriptomics, proteomics, metabolomics, methylomics, 
genetic_variations, dna_binding_sites. 
Default: transcriptomics.}

\item{user_name}{The name of the user who uploaded the signatures.}

\item{signature_name}{The name of the signatures to be looked up by.}

\item{organism}{The organism to be looked up by.}

\item{phenotype}{The phenotype to be looked up by.}

\item{sample_type}{The sample type to be looked up by.}
}
\description{
Get a list of signatures available in the database
}
\examples{

# Create a db connection
conn_handler <- SigRepo::newConnHandler(
 dbname = "sigrepo", 
 host = "montilab.bu.edu", 
 port = 3306, 
 user = "guest", 
 password = "guest"
)

# Get a list of signatures available in the database
signature_list <- sigRepo::searchSignatures(conn_handler = conn_handler)

}
