% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchUser.R
\name{searchUser}
\alias{searchUser}
\title{searchUser}
\usage{
searchUser(conn_handler, user_name = NULL, verbose = TRUE)
}
\arguments{
\item{conn_handler}{A handler uses to establish connection to the database 
obtained from SigRepo::newConnhandler() (required)}

\item{user_name}{A list of user names to search by. Default is NULL which
will return all of the users in the database}

\item{verbose}{A logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Get phenotypes in the database
}
