% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleQuote.R
\name{singleQuote}
\alias{singleQuote}
\title{singleQuote}
\usage{
singleQuote(myString)
}
\arguments{
\item{myString}{the string to add single quotation marks to}
}
\value{
a string, flanked by single quotes
}
\description{
Add single quotation marks around a string
}
\examples{
singleQuote("String to put in single quotes")
}
