% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFindingQuery.R
\name{sqlFindingQuery}
\alias{sqlFindingQuery}
\title{sqlFindingQuery}
\usage{
sqlFindingQuery(
  dbTable,
  fields = c("*"),
  jsonFields = NULL,
  ins = NULL,
  betweens = NULL,
  orderFields = NULL,
  distinct = TRUE,
  verbose = FALSE,
  collapse_by = " AND ",
  handle = newConnHandle(),
  disconnectAfter = TRUE
)
}
\arguments{
\item{dbTable}{string; the table to select from}

\item{fields}{character vector, the fields to select from the target table}

\item{jsonFields}{named list. name is the field with the json. values are the
names of the keys in that field you're trying to get values of.}

\item{ins}{named list where the names are the fields to narrow search by
and the values are vectors of the values to look for in those fields}

\item{betweens}{named list where the names are the fields to narrow search by
and the values are vectors of length 2 containing the endpoints of the
range to consider in those fields}

\item{orderFields}{a vector of fields to order the query results by}

\item{distinct}{a boolean indicating whether to return unique results}

\item{verbose}{default to be FALSE. set TRUE to print messages}

\item{collapse_by}{how to collapse the input queries, default to be "AND"}

\item{handle}{DB connection handle}

\item{disconnectAfter}{whether disconnect the DB after finding. default to be TRUE}
}
\description{
Constructs sql query based on where clause, if one is needed,
  and executes final query as output#'
}
