% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlGeneric.R
\name{sqlGeneric}
\alias{sqlGeneric}
\title{sqlGeneric}
\usage{
sqlGeneric(query, conn = newConnHandle(), disconnectAfter = T)
}
\arguments{
\item{query}{the query to submit. String}

\item{conn}{connection Handle. uses default newConnHandle call if none
defined.}

\item{disconnectAfter}{Boolean: whether to disconnect the handle after
executing the query}
}
\description{
Submit a query to the database
}
