% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFunctions.R
\name{update_table_sql}
\alias{update_table_sql}
\title{update_table_sql}
\usage{
update_table_sql(
  conn,
  table,
  update_coln_var,
  update_coln_val,
  filter_coln_var,
  filter_coln_val
)
}
\arguments{
\item{conn}{An established database connection using newConnhandler()}

\item{table}{A table in the database}

\item{update_coln_var}{A list of column variables in the given table}

\item{update_coln_val}{A list of column values associated with the column variables}

\item{filter_coln_var}{A list of column values associated with the column variables}

\item{filter_coln_val}{A list of column values associated with the column variables}
}
\description{
update a table into the database
}
