% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSignature.R
\name{updateSignature}
\alias{updateSignature}
\title{updateSignature}
\usage{
updateSignature(conn_handler, signature_id, omic_signature)
}
\arguments{
\item{conn_handler}{An established connection handler using SigRepo::newConnhandler()}

\item{signature_id}{a unique signature id in the database that needs to be updated}

\item{omic_signature}{An R6 class object from the OmicSignature package}
}
\description{
Update a signature in the database
}
\examples{

# Create a db connection
conn_handler <- SigRepo::newConnHandler(
 dbname = "sigrepo", 
 host = "montilab.bu.edu", 
 port = 3306, 
 user = "guest", 
 password = "guest"
)

# Get a list of signatures that belongs to user = 'guest'
signature_tbl <- sigRepo::searchSignature(
 conn_handler = conn_handler,
 user_name = "guest"
)

# Update the desired signature (NOT RUN)
# SigRepo::updateSignature(
#  conn_handler = conn_handler,
#  signature_id = signature_tbl$signature_id,
#  omic_signature = omic_signature 
# )

}
