% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUser.R
\name{updateUser}
\alias{updateUser}
\title{updateUser}
\usage{
updateUser(
  conn_handler,
  user_name,
  password = NULL,
  email = NULL,
  first_name = NULL,
  last_name = NULL,
  affiliation = NULL,
  role = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{Name of a user to be updated (required).}

\item{password}{Password of a user to be updated. Default is NULL.}

\item{email}{Email of a user to be updated. Default is NULL.}

\item{first_name}{First name of a user to be updated. Default is NULL}

\item{last_name}{Last name of a user to be updated. Default is NULL.}

\item{affiliation}{Affiliation of the user. Default is NULL.}

\item{role}{Role of a user to be updated. Choices are admin/editor/viewer.}

\item{verbose}{a logical value indicates whether or not to print the
diagnostic messages. Default is \code{TRUE}.}
}
\description{
Update a user information in the database
}
\examples{

 # Establish a Connection Handler using newConnHandler if not done so already.

# SigRepo::updateUser(
# conn_handler = conn,
# user_name = "test_user", # required
# password = "new_password", 
# email = "test_email",
# affiliation = "test_affiliation",
# last_name = "test_last_name",
# role = "editor",
# verbose = FALSE
# )


}
