% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUser.R
\name{updateUser}
\alias{updateUser}
\title{updateUser}
\usage{
updateUser(
  conn_handler,
  user_name,
  password = NULL,
  email = NULL,
  first_name = NULL,
  last_name = NULL,
  affiliation = NULL,
  role = NULL
)
}
\arguments{
\item{conn_handler}{An established database connection using SigRepo::newConnhandler()}

\item{user_name}{Name of a user to be updated (required).}

\item{password}{Password of a user to be updated. Default is NULL.}

\item{email}{Email of a user to be updated. Default is NULL.}

\item{last_name}{Last name of a user to be updated. Default is NULL.}

\item{affiliation}{First name of a user to be updated. Default is NULL.}

\item{role}{Role of a user to be updated. Choices are admin/editor/viewer.}
}
\description{
Update a user information in the database
}
