% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSignatureFile.R
\name{writeSignatureFile}
\alias{writeSignatureFile}
\title{writeSignatureFile}
\usage{
writeSignatureFile(uploadPath, fileObject, fileName)
}
\arguments{
\item{uploadPath}{path to upload / write the signature file}

\item{fileObject}{signature object}

\item{fileName}{name of the file}
}
\description{
writeSignatureFile
}
